+++ changelog.md
@@ -2,5 +2,6 @@
 
 ## 3.7.0
 * Port to 1.19.2
-* 
+* Ender Dragon
+  * Reduced bonus damage (~~+240%~~ -> +225% at max difficulty)
 * Fixed generating a bak config file due to wrong explosion config option